draft post component and edit post bug fix
